@@ -1,63 +1,134 @@
 {
   "name": "@mongodb-js/compass-field-store",
   "productName": "Compass Compass Field Store package",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "FieldStore keeps track of available fields in a collection.",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "ci": "npm run check && npm test",
+    "clean": "rimraf lib",
+    "precompile": "npm run clean",
+    "compile": "cross-env NODE_ENV=production webpack --config ./config/webpack.prod.config.js --progress",
+    "start": "webpack-dev-server --config ./config/webpack.dev.config.js",
+    "start:watch": "npm run clean && webpack --config ./config/webpack.watch.config.js",
+    "test": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\"",
+    "test:watch": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\" --watch",
+    "test:karma": "cross-env NODE_ENV=test karma start",
+    "cover": "nyc npm run test",
+    "ci": "npm run cover && npm run test:karma && npm run compile && npm run check",
     "fmt": "mongodb-js-fmt ./*.js ./test/*.js",
-    "check": "mongodb-js-precommit './src/**/*{.js,.jsx}' './test/**/*.js' index.js",
-    "compile": "babel ./src --out-dir ./lib",
+    "check": "mongodb-js-precommit './src/**/*{.js,.jsx}' './test/**/*.js' './electron/**/*.js' './config/**/*{.js,.jsx}'",
+    "link-plugin": "./scripts/link.sh",
+    "unlink-plugin": "./scripts/unlink.sh",
     "prepublish": "npm run compile",
-    "pretest": "npm run compile",
-    "prestart": "npm run compile",
-    "start": "electron ./electron",
-    "test": "npm run compile && mocha test"
+    "storybook": "start-storybook -p 9001 -c .storybook"
   },
   "precommit": [
     "check"
   ],
+  "dependency-check": {
+    "ignore": [
+      "react-dom",
+      "debug"
+    ]
+  },
   "license": "Apache-2.0",
+  "peerDependencies": {
+    "debug": "^3.0.1",
+    "mongodb-schema": "^8.0.0"
+  },
   "devDependencies": {
+    "@storybook/react": "^3.2.0",
+    "autoprefixer": "^9.4.6",
     "babel-cli": "^6.10.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.1.1",
+    "babel-minify": "^0.2.0",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-syntax-object-rest-spread": "^6.13.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.11.1",
+    "babel-preset-minify": "^0.2.0",
+    "babel-preset-react": "^6.24.1",
     "babel-register": "^6.9.0",
     "bootstrap": "https://github.com/twbs/bootstrap/archive/v3.3.5.tar.gz",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^5.1.0",
-    "chai-enzyme": "^0.5.0",
-    "debug": "^2.2.0",
-    "enzyme": "^2.4.1",
+    "chai": "^4.1.2",
+    "chai-enzyme": "1.0.0-beta.0",
+    "classnames": "^2.2.5",
+    "cross-env": "^5.0.1",
+    "css-loader": "^0.28.1",
+    "debug": "^3.0.1",
+    "electron": "^3.0.7",
+    "electron-packager": "^8.7.0",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
     "eslint-config-mongodb-js": "^2.1.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^3.0.1",
     "font-awesome": "^4.7.0",
-    "hadron-app": "^2.0.0",
-    "hadron-app-registry": "^7.0.0",
-    "hadron-react-components": "^1.1.0",
-    "hadron-reflux-store": "0.0.3",
-    "hadron-spectron": "^0.0.1",
-    "hadron-type-checker": "^2.0.0",
-    "jsdom": "^9.4.2",
-    "mocha": "^2.5.3",
-    "mock-require": "^2.0.2",
-    "mongodb-data-service": "^11.0.3",
+    "hadron-app": "^3.1.0",
+    "hadron-app-registry": "^7.1.0",
+    "html-webpack-plugin": "^3.2.0",
+    "ignore-loader": "^0.1.2",
+    "istanbul-instrumenter-loader": "^3.0.0",
+    "jsdom": "^11.1.0",
+    "jsdom-global": "^3.0.2",
+    "karma": "^1.7.0",
+    "karma-chai": "^0.1.0",
+    "karma-chai-sinon": "^0.1.5",
+    "karma-electron": "^5.2.1",
+    "karma-mocha": "^1.3.0",
+    "karma-mocha-reporter": "^2.2.4",
+    "karma-sinon": "^1.0.5",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^4.0.0-rc.5",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.5",
+    "lodash.clonedeep": "^4.5.0",
+    "lodash.includes": "^4.3.0",
+    "lodash.isarray": "^4.0.0",
+    "lodash.isnumber": "^3.0.3",
+    "lodash.isstring": "^4.0.1",
+    "lodash.mergewith": "^4.6.1",
+    "lodash.pick": "^4.4.0",
+    "lodash.union": "^4.6.0",
+    "lodash.uniq": "^4.5.0",
+    "mocha": "^4.1.0",
+    "mocha-webpack": "^2.0.0-beta.0",
+    "mock-require": "^3.0.3",
+    "mongodb-data-service": "^12.4.17",
     "mongodb-js-fmt": "^0.0.3",
-    "mongodb-js-precommit": "^0.2.8",
+    "mongodb-js-precommit": "^1.1.0",
+    "mongodb-reflux-store": "0.0.1",
+    "mongodb-schema": "^8.0.0",
+    "node-loader": "^0.6.0",
+    "nyc": "^13.1.0",
+    "peer-deps-externals-webpack-plugin": "^1.0.2",
+    "postcss-loader": "^2.0.6",
     "pre-commit": "^1.1.3",
     "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-fontawesome": "^1.6.1",
-    "reflux": "^0.4.1",
-    "sinon": "^1.17.6",
-    "sinon-chai": "^2.8.0",
-    "watch": "^1.0.2"
-  },
-  "dependencies": {
-    "lodash": "^4.17.4",
-    "mongodb-schema": "^8.0.0",
-    "reflux-state-mixin": "^0.7.0"
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
+    "react-hot-loader": "^4.1.0",
+    "react-redux": "^5.0.6",
+    "react-storybook-addon-chapters": "^1.1.0",
+    "redux": "^3.7.2",
+    "reflux-state-mixin": "^0.8.0",
+    "resolve": "^1.4.0",
+    "rimraf": "^2.6.1",
+    "semver": "^5.4.1",
+    "shebang-loader": "0.0.1",
+    "sinon": "^3.2.0",
+    "sinon-chai": "^2.11.0",
+    "style-loader": "^0.18.2",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.29.0",
+    "webpack-bundle-analyzer": "^3.0.3",
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.1.14",
+    "webpack-merge": "^4.2.1",
+    "webpack-node-externals": "^1.7.2",
+    "webpack-sources": "1.0.1"
   }
 }
